requirements.txt
pyspark
request
feedparser
